+++ examples/number_range.py
@@ -2,7 +2,6 @@ import asyncio
 
 from async_graph_data_flow import AsyncExecutor, AsyncGraph
 
-
 BUFFER = 3
 count = 0
 batch_list = []
@@ -46,7 +45,7 @@ async def load(data):
 
 if __name__ == "__main__":
     etl_graph = AsyncGraph()
-    etl_graph.add_node(extract, name="extract", max_tasks=1, queue_size=1_000)
+    etl_graph.add_node(extract, name="extract", max_tasks=1)
     etl_graph.add_node(transform)
     etl_graph.add_node(load)
     etl_graph.add_edge("extract", "transform")